requirements.txt
# Runtime deps for Hugging Face Spaces (avoid strict pins that conflict with gradio[oauth])
huggingface_hub>=0.33.0,<1.0.0
transformers>=4.45.0,<5.0.0
torch>=2.2.0,<2.7.0
safetensors>=0.4.0,<1.0.0

# Scraping / parsing / feeds
feedparser>=6.0.11,<7.0.0
trafilatura>=1.12.0,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
lxml>=5.3.0,<7.0.0
readability-lxml>=0.8.1,<1.0.0
googlenewsdecoder>=0.1.7,<1.0.0
requests>=2.32.3,<3.0.0
pandas>=2.2.0,<3.0.0
